SearchableCombobox getLabels() and onSelect() changes
If the parameter getLabels() is missing, then it is defined
as a function that returns its arguments.

getLabels = (options) => options;

The options argument is supplied at invocation by the useCombobox hook.

The onSelect() handler function is now provided with the ability to reset the
combobox by means of handing it over the `inputValue` state setter as the second
argument.

example:
onSelect(inputValue, setInputValue)